backend/pyproject.toml
[project]
name = "campgear-backend"
version = "0.1.0"
description = "CampGear Store — Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic[email]>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "passlib[bcrypt]>=1.7,<2",
    "bcrypt>=4.0,<4.1",
    "python-jose[cryptography]>=3.3,<4",
    "python-multipart>=0.0.18",
    "pillow>=11,<12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.28,<1",
    "ruff>=0.9,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "S", "B"]
ignore = ["S101"]  # allow assert in tests

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
